+++ packages/assets/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.43.0](https://github.com/randytarampi/me/compare/v0.42.2...v0.43.0) (2018-10-16)
+
+**Note:** Version bump only for package @randy.tarampi/assets
+
+
+
+
+
 ## [0.42.2](https://github.com/randytarampi/me/compare/v0.42.1...v0.42.2) (2018-10-16)
 
 **Note:** Version bump only for package @randy.tarampi/assets